Controller: I re-wrote most of the controller handling. It was spiraling out of control due to some poor planning, creating at least 2 extra jobs for most steps. This led to a huge backlog for processing in APT.
I added in an active job dict in the controller to more easily identify jobs, removing most of the for loops in the controller, should save a lot of unnecessary looping. The dict keys are active jobs, value is a list, v[0] is the active processing target, v[1] is the origin.

I also added in proper checking to jobs so as to not create a bunch of redundant jobs that APT can't possibly keep up with.

Dispatcher:
I also re-wrote basically the entire dispatcher loop, it was unwieldly, I added in additional deques, one per monitor, and simplified the logic to make it easier to work with. I have remove most/all of the locks.. not sure how that will go yet being new to threading, but I think that it will work out ok since the deque will only be accessed by single endpoints on each side.